# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Travel Risk API",
    "description": "Backend of the tourism travel safety risk assessment platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/config": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "Get Active Config",
        "operationId": "get_active_config_api_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "config"
        ],
        "summary": "Create Config",
        "operationId": "create_config_api_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/config/versions/{version}/activate": {
      "post": {
        "tags": [
          "config"
        ],
        "summary": "Activate Config",
        "operationId": "activate_config_api_config_versions__version__activate_post",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/config/curves": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "Config Curves",
        "operationId": "config_curves_api_config_curves_get",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Config Curves Api Config Curves Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/evaluations": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Create Evaluation",
        "operationId": "create_evaluation_api_evaluations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Evaluation Api Evaluations Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "List Evaluations",
        "operationId": "list_evaluations_api_evaluations_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Evaluations Api Evaluations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations/{evaluation_id}": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Evaluation",
        "operationId": "get_evaluation_api_evaluations__evaluation_id__get",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Evaluation Api Evaluations  Evaluation Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "evaluations"
        ],
        "summary": "Delete Evaluation",
        "operationId": "delete_evaluation_api_evaluations__evaluation_id__delete",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations/{evaluation_id}/regions/{region_id}/individuals": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "List Individuals",
        "operationId": "list_individuals_api_evaluations__evaluation_id__regions__region_id__individuals_get",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Evaluation Id"
            }
          },
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Region Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Individuals Api Evaluations  Evaluation Id  Regions  Region Id  Individuals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations/{evaluation_id}/export": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Export Evaluation",
        "operationId": "export_evaluation_api_evaluations__evaluation_id__export_get",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Evaluation Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(xlsx|pdf)$",
              "default": "xlsx",
              "title": "Format"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(uk|en)$",
              "default": "uk",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChiScaleParams": {
        "properties": {
          "L": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "L"
          },
          "BA": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Ba"
          },
          "A": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "A"
          },
          "AA": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Aa"
          },
          "H": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "H"
          }
        },
        "type": "object",
        "required": [
          "L",
          "BA",
          "A",
          "AA",
          "H"
        ],
        "title": "ChiScaleParams"
      },
      "ConeParams": {
        "properties": {
          "base": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Base"
          },
          "scale": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Scale"
          }
        },
        "type": "object",
        "required": [
          "base",
          "scale"
        ],
        "title": "ConeParams"
      },
      "ConfigCreate": {
        "properties": {
          "params": {
            "$ref": "#/components/schemas/ConfigParams"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "params"
        ],
        "title": "ConfigCreate"
      },
      "ConfigOut": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "version",
          "created_at",
          "active",
          "comment",
          "params"
        ],
        "title": "ConfigOut"
      },
      "ConfigParams": {
        "properties": {
          "term_multipliers": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Term Multipliers"
          },
          "chi_scale": {
            "$ref": "#/components/schemas/ChiScaleParams"
          },
          "z_spline": {
            "$ref": "#/components/schemas/ZSplineParams"
          },
          "cone": {
            "$ref": "#/components/schemas/ConeParams"
          },
          "fuzzification_boundaries": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 6,
            "minItems": 6,
            "title": "Fuzzification Boundaries"
          },
          "risk_thresholds": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Risk Thresholds"
          }
        },
        "type": "object",
        "required": [
          "term_multipliers",
          "chi_scale",
          "z_spline",
          "cone",
          "fuzzification_boundaries",
          "risk_thresholds"
        ],
        "title": "ConfigParams"
      },
...
      "EvaluationCreate": {
        "properties": {
          "region_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Ids"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "title": "EvaluationCreate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ZSplineParams": {
        "properties": {
          "a": {
            "type": "number",
            "title": "A"
          },
          "b": {
            "type": "number",
            "title": "B"
          }
        },
        "type": "object",
        "required": [
          "a",
          "b"
        ],
        "title": "ZSplineParams"
      }
    }
  }
}